source:trunk/.irbrc  * Preload 'apache/fakerequest' to emulate a mod_ruby environment
source:trunk/Rakefile
 * Added an install_dependencies task

source:trunk/spec/templatefactory_spec.rb
source:trunk/spec/broker_spec.rb
source:trunk/spec/cookieset_spec.rb
source:trunk/spec/applet_spec.rb
source:trunk/spec/activerecordstore_spec.rb
 * Aesthetic cleanup
 * Preload 'apache/fakerequest' to emulate a mod_ruby environment for
   autotest, etc.

source:trunk/spec/transaction_spec.rb
source:trunk/spec/formvalidator_spec.rb
source:trunk/tests/transaction.tests.rb
source:trunk/tests/formvalidator.tests.rb
 * Converted Test::Unit tests to RSpec examples

source:trunk/lib/arrow/config.rb
source:trunk/spec/dispatcher_spec.rb
 * Some Ruby 1.8.6 taint-model fixes
 * Converted remainder of Test::Unit tests to RSpec specifications

source:trunk/spec/lib/helpers.rb
source:trunk/spec/lib/matchers.rb
 * Added the beginnings of some RSpec helper/matcher modules

source:trunk/spec/TEMPLATE.rb.tpl
 * Converted to RSpec >= 1.0.x style specs

source:trunk/utils.rb
 * Preload RubyGems before testing for installed libraries

source:trunk/lib/apache/fakerequest.rb
 * Move the Apache::Request method list into the mock environment instead of
   living in the real transaction class.

source:trunk/lib/arrow/dispatcher.rb
source:trunk/lib/arrow.rb
 * Preload RubyGems
 * COMPAT: Changed the RubyChildInitHandler into a dedicated class that
   handles the configfile loading. This will require changing any
   'RubyChildInitHandler' directives you may have in your Apache configs

source:trunk/lib/arrow/session/activerecordstore.rb
 * Squelch ActiveRecord errors while saving the session if/until we can
   figure out database locking

source:trunk/lib/arrow/session/filelock.rb
 * Fix a bug which happened when the lockfile was removed before the session
   became unlocked.

source:trunk/lib/arrow/transaction.rb
 * Separated out the handler status into a separate attribute. Not sure if
   this is necessary or not quite yet.

source:trunk/lib/arrow/formvalidator.rb
 * Aliased #has_errors? to #errors?
 * Tweaked the human-readable error descriptions a bit

source:trunk/lib/arrow/template.rb
 * Undid some debug log spam
 * Added some early detection for some error conditions in adding
   definition sets

source:trunk/lib/arrow/testcase.rb
 * Preload RubyGems

source:trunk/install.rb
 * Updated dependency URLs

--HG--
extra : convert_revision : svn%3Ab4f3b83b-18d6-0310-a2de-c8f40de7a630/trunk%40403

+++ lib/arrow/session/filelock.rb
@@ -129,7 +129,7 @@ class Arrow::Session::FileLock < Arrow::Session::Lock
 			path = @lockfile.path
 			@lockfile.close
 			@lockfile = nil
-			File.delete( path.untaint )
+			File.delete( path.untaint ) if File.exist?( path.untaint )
 		end
 	end
